imple new async debug rpcs
