+++ Cargo.toml
@@ -59,7 +59,9 @@ members = [
     "crates/rpc/rpc-cfx-types", 
     "crates/rpc/rpc-eth-api",
     "crates/cfxcore/parity-trace-types",
-    "crates/rpc/rpc", "crates/util/util-macros",
+    "crates/rpc/rpc",
+    "crates/util/util-macros",
+    "crates/rpc/rpc-utils",
 ]
 
 resolver = "2"
@@ -151,40 +153,46 @@ cfx-rpc-cfx-types = { path = "./crates/rpc/rpc-cfx-types" }
 cfx-parity-trace-types = { path = "./crates/cfxcore/parity-trace-types" }
 cfx-rpc-eth-api = { path = "./crates/rpc/rpc-eth-api" }
 cfx-util-macros = { path = "./crates/util/util-macros" }
+cfx-rpc = { path = "./crates/rpc/rpc" }
+cfx-rpc-utils = { path = "./crates/rpc/rpc-utils" }
 
 serde = { version = "1.0", features = ["derive", "alloc"] }
 serde_json = "1.0"
 serde_derive = "1.0"
-fail = "0.4.0"
-thiserror = "1.0.63"
-anyhow = "1.0"
-chrono = "=0.4.38"
-error-chain = { version = "0.12" }
-jsonrpc-core = { version = "15.1.0" }
-log = "0.4"
-strum = "0.20"
-strum_macros = "0.20"
-lazy_static = "1.4"
-parking_lot = "0.11"
-rayon = "1.10"
-jsonrpsee = "0.24.4"
-once_cell = "1.17.1"
-rand = "0.7"
-itertools = "0.10.0"
-
-byteorder = "1.2.7"
 rlp = "0.4.0"
 rlp_derive = { git = "https://github.com/Conflux-Chain/conflux-parity-deps.git", rev = "1597a9cab02343eb2322ca0ac58d39b64e3f42d1"  }
 hex = "0.4.3"
 rustc-hex = "2.1"
 
+fail = "0.4.0"
+thiserror = "1.0.63"
+anyhow = "1.0"
+error-chain = { version = "0.12" }
+strum = "0.20"
+strum_macros = "0.20"
+
+# parallelism
+parking_lot = "0.11"
+rayon = "1.10"
+
+# alloy & revm
 alloy-sol-types = "0.7.2"
 alloy-primitives = "0.7.2"
 alloy-rpc-types = { git = "https://github.com/alloy-rs/alloy", rev = "4e22b9e" }
 alloy-rpc-types-trace = { git = "https://github.com/alloy-rs/alloy", rev = "4e22b9e" }
 revm = "8.0"
 
-bls-signatures = { git = "https://github.com/Conflux-Chain/bls-signatures.git", rev = "fb52187df92d27c365642cb7e7b2aaf60437cf9c", default-features = false, features = ["multicore", "blst"] }
+# jsonrpc framework(old)
+jsonrpc-core = "15.1.0"
+jsonrpc-tcp-server = "15.1.0"
+jsonrpc-http-server = "15.1.0"
+jsonrpc-derive = "15.1.0"
+jsonrpc-core-client = "15.1.0"
+jsonrpc-pubsub = "15.1.0"
+jsonrpc-ws-server = "15.1.0"
+
+# jsonrpc framework(new)
+jsonrpsee = "0.24.4"
 
 # async
 async-trait = "0.1"
@@ -193,6 +201,14 @@ async-trait = "0.1"
 fixed-hash = "0.5"
 keccak-hash = "0.5"
 tiny-keccak = "2.0.2"
+bls-signatures = { git = "https://github.com/Conflux-Chain/bls-signatures.git", rev = "fb52187df92d27c365642cb7e7b2aaf60437cf9c", default-features = false, features = ["multicore", "blst"] }
 
 # misc
 clap = "2"
+rand = "0.7"
+log = "0.4"
+lazy_static = "1.4"
+itertools = "0.10.0"
+once_cell = "1.17.1"
+chrono = "=0.4.38"
+byteorder = "1.2.7"
