+++ bins/conflux/Cargo.toml
@@ -26,10 +26,10 @@ rpassword = "4.0.1"
 io = { workspace = true }
 runtime = { workspace = true }
 ctrlc = { git = "https://github.com/paritytech/rust-ctrlc.git", rev="b523017108bb2d571a7a69bd97bc406e63bc7a9d" }
-jsonrpc-core = "15.1.0"
-jsonrpc-tcp-server = "15.1.0"
-jsonrpc-http-server = "15.1.0"
-jsonrpc-core-client = { version = "15.1.0", features = ["http"] }
+jsonrpc-core = { workspace = true }
+jsonrpc-tcp-server = { workspace = true }
+jsonrpc-http-server = { workspace = true }
+jsonrpc-core-client = { workspace = true, features = ["http"] }
 error-chain = { version = "0.12" }
 log = "0.4"
 slab = "0.4"
