+++ crates/rpc/rpc-utils/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "cfx-rpc-utils"
+edition = "2021"
+version.workspace = true
+authors.workspace = true
+description.workspace = true
+documentation.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+repository.workspace = true
+license-file.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+jsonrpsee = { workspace = true }
+jsonrpc-core = { workspace = true}
