+++ crates/rpc/rpc/src/lib.rs
@@ -1,7 +1,9 @@
+mod debug;
 mod net;
 mod rpc;
 mod web3;
 
+pub use debug::DebugApi;
 pub use net::NetApi;
 pub use rpc::RPCApi;
 pub use web3::Web3Api;
